Updated to Fusion SDK 1.0.5 with the small bugfix to file/CELFileData.h mentioned a couple of commits ago. Made changes to Makevars, Makevars.win in order to build a minimal library Added target: dependency to _Makefile, useful for updating the SDK. Added fusion2affxparser.sh which trims the SDK. Removed src/old and some README, CHANGES files that were old and out of date. Replaced all of this with inst/info/README.
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/affxparser@18117 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/fusion_sdk/calvin_files/utils/src/AffymetrixGuid.cpp
@@ -44,14 +44,14 @@ using namespace affymetrix_calvin_utilities;
 void InitializeRandomNumberGenerator()
 {
 	// Get the current time as a starting point for a seed.
-	u_int64_t seed = (u_int64_t) time(NULL);
+	unsigned int seed = (unsigned int) time(NULL);
 
 	// Multiply the time by its address and the process id.
-	seed *= (u_int64_t)(&seed);
-	seed *= (u_int64_t)getpid();
+	seed *= (unsigned int)(&seed);
+	seed *= getpid();
 
 	// Initialize the random number generator.
-	srand((unsigned int)seed);
+	srand(seed);
 }
 
 /*
